+++ lib/drivers/index.ts
@@ -1,7 +1,0 @@
-export { Driver } from './Driver';
-export { JSONDriver as JSON } from './JSON';
-export { YAML } from './YAML';
-export { BSON } from './BSON';
-export { TOML } from './TOML';
-export { INI } from './INI';
-export { JSON5 } from './JSON5';