+++ lib/index.ts
@@ -1,5 +1,4 @@
 import Database from './database/Database';
-const { Drivers } = Database;
 
-export { Database, Drivers };
+export { Database };
 export default Database;