+++ lib/interfaces/DatabaseMap.ts
@@ -1,3 +1,0 @@
-export interface DatabaseMap {
-  [key: string]: any;
-};