+++ lib/interfaces/DatabaseSignature.ts
@@ -1,3 +1,0 @@
-export type DatabaseSignature<V> = {
-  [key in keyof V]: any;
-};